+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.15.0 - Unreleased
 ### Added
 - Added site selector to campaign and mailing list reports ([#125](https://github.com/putyourlightson/craft-campaign/issues/125)).
+- Added the ability to import contacts to multiple lists in one go ([#126](https://github.com/putyourlightson/craft-campaign/issues/126)).
 
 ### Fixed
 - Fixed an error that could appear if a campaign type's or mailing list type's site was deleted.